move : Coupon 도메인 Test 분리
