+++ src/test/java/bc1/gream/domain/user/service/CouponServiceTest.java
@@ -6,9 +6,9 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import bc1.gream.domain.user.entity.Coupon;
+import bc1.gream.domain.user.coupon.entity.Coupon;
+import bc1.gream.domain.user.coupon.repository.CouponRepository;
 import bc1.gream.domain.user.entity.User;
-import bc1.gream.domain.user.repository.CouponRepository;
 import bc1.gream.test.CouponTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
