+++ src/test/java/bc1/gream/test/BaseIntegrationTest.java
@@ -6,9 +6,9 @@ import bc1.gream.domain.order.repository.OrderRepository;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.repository.ProductRepository;
 import bc1.gream.domain.sell.repository.SellRepository;
-import bc1.gream.domain.user.entity.Coupon;
+import bc1.gream.domain.user.coupon.entity.Coupon;
+import bc1.gream.domain.user.coupon.repository.CouponRepository;
 import bc1.gream.domain.user.entity.User;
-import bc1.gream.domain.user.repository.CouponRepository;
 import bc1.gream.domain.user.repository.UserRepository;
 import bc1.gream.global.security.WithMockCustomUser;
 import org.junit.jupiter.api.Disabled;
